# Conflicts:
#	Cargo.toml

+++ src/naming/cluster/mod.rs
@@ -1,11 +1,5 @@
 //distor cluster
 
-use std::{
-    collections::{HashMap, HashSet},
-    convert::TryFrom,
-    sync::Arc,
-};
-
 use self::{
     model::{
         NamingRouteRequest, NamingRouterResponse, ProcessRange, SnapshotDataInfo,
@@ -14,12 +8,21 @@ use self::{
     node_manage::{NodeManageRequest, NodeManageResponse},
 };
 use crate::common::constant::GRPC_HEAD_KEY_CLUSTER_ID;
+use crate::mcp::model::actor_model::{McpManagerReq, McpManagerResult};
+use crate::mcp::model::sse_model::SseStreamManageAsyncCmd;
+use crate::mcp::sse_manage::SseConnUtils;
 use crate::metrics::model::{MetricsRequest, MetricsResponse};
 use crate::naming::cluster::model::SnapshotForSend;
 use crate::naming::model::{DistroData, Instance};
 use crate::{
     common::appdata::AppShareData,
     naming::core::{NamingCmd, NamingResult},
+    openapi,
+};
+use std::{
+    collections::{HashMap, HashSet},
+    convert::TryFrom,
+    sync::Arc,
 };
 
 pub mod instance_delay_notify;
@@ -44,6 +47,8 @@ pub async fn handle_naming_route(
     req: NamingRouteRequest,
     extend_info: HashMap<String, String>,
 ) -> anyhow::Result<NamingRouterResponse> {
+    #[cfg(feature = "debug")]
+    log::info!("handle_naming_route req:{:?}", &req);
     match req {
         NamingRouteRequest::Ping(cluster_id) => {
             //更新node_id节点活跃状态
@@ -66,11 +71,13 @@ pub async fn handle_naming_route(
         NamingRouteRequest::SyncUpdateInstance { mut instance } => {
             let cluster_id = get_cluster_id(extend_info)?;
             reset_cluster_info(cluster_id, &mut instance);
-            app.naming_inner_node_manage
-                .do_send(NodeManageRequest::AddClientId(
-                    cluster_id,
-                    instance.client_id.clone(),
-                ));
+            if instance.from_cluster == cluster_id {
+                app.naming_inner_node_manage
+                    .do_send(NodeManageRequest::AddClientId(
+                        cluster_id,
+                        instance.client_id.clone(),
+                    ));
+            }
             let cmd = NamingCmd::Update(instance, None);
             let _: NamingResult = app.naming_addr.send(cmd).await??;
         }
@@ -89,19 +96,27 @@ pub async fn handle_naming_route(
             let mut client_sets = HashSet::new();
             for instance in &mut batch_receive.update_instances {
                 reset_cluster_info(cluster_id, instance);
-                client_sets.insert(instance.client_id.clone());
+                if instance.from_cluster == cluster_id {
+                    client_sets.insert(instance.client_id.clone());
+                }
             }
             /*
             for instance in &mut batch_receive.remove_instances {
                 reset_cluster_info(cluster_id, instance);
             }
              */
-            app.naming_inner_node_manage
-                .do_send(NodeManageRequest::AddClientIds(cluster_id, client_sets));
-            app.naming_addr
-                .do_send(NamingCmd::DeleteBatch(batch_receive.remove_instances));
-            app.naming_addr
-                .do_send(NamingCmd::UpdateBatch(batch_receive.update_instances));
+            if !client_sets.is_empty() {
+                app.naming_inner_node_manage
+                    .do_send(NodeManageRequest::AddClientIds(cluster_id, client_sets));
+            }
+            if !batch_receive.remove_instances.is_empty() {
+                app.naming_addr
+                    .do_send(NamingCmd::DeleteBatch(batch_receive.remove_instances));
+            }
+            if !batch_receive.update_instances.is_empty() {
+                app.naming_addr
+                    .do_send(NamingCmd::UpdateBatch(batch_receive.update_instances));
+            }
         }
         NamingRouteRequest::RemoveClientId { client_id } => {
             app.naming_inner_node_manage
@@ -138,10 +153,14 @@ pub async fn handle_naming_route(
             let mut client_sets = HashSet::new();
             for instance in &mut snapshot_receive.instances {
                 reset_cluster_info(cluster_id, instance);
-                client_sets.insert(instance.client_id.clone());
+                if instance.from_cluster == cluster_id {
+                    client_sets.insert(instance.client_id.clone());
+                }
             }
-            app.naming_inner_node_manage
-                .do_send(NodeManageRequest::AddClientIds(cluster_id, client_sets));
+            if !client_sets.is_empty() {
+                app.naming_inner_node_manage
+                    .do_send(NodeManageRequest::AddClientIds(cluster_id, client_sets));
+            }
             //增量数据
             app.naming_addr
                 .do_send(NamingCmd::ReceiveSnapshot(snapshot_receive));
@@ -206,6 +225,56 @@ pub async fn handle_naming_route(
                     let cmd = NodeManageRequest::SendSnapshot(cluster_id, snapshot);
                     app.naming_inner_node_manage.send(cmd).await??;
                 }
+            }
+        }
+        NamingRouteRequest::QueryServiceSubscriberPage(param) => {
+            let resp = app
+                .naming_addr
+                .send(NamingCmd::QueryServiceSubscribersPageV2(param))
+                .await??;
+            if let NamingResult::ServiceSubscribersPage((total, list)) = resp {
+                return Ok(NamingRouterResponse::ServiceSubscribersPage((total, list)));
+            }
+        }
+        NamingRouteRequest::McpMessages {
+            server_key,
+            session_id,
+            request,
+            headers,
+        } => {
+            let mcp_server = if let Ok(Ok(McpManagerResult::ServerInfo(Some(server)))) = app
+                .mcp_manager
+                .send(McpManagerReq::GetServerByKey(server_key.clone()))
+                .await
+            {
+                server
+            } else {
+                return Err(anyhow::anyhow!("McpServer not found"));
+            };
+            let mut ref_headers = HashMap::new();
+            for (key, value) in headers.iter() {
+                ref_headers.insert(key.as_str(), value.as_bytes());
+            }
+            match openapi::mcp::api::handle_request(
+                app,
+                request,
+                &mcp_server,
+                session_id.as_ref(),
+                ref_headers,
+            )
+            .await
+            {
+                Ok(result) => {
+                    let message = SseConnUtils::create_sse_message(&result);
+                    app.sse_stream_manager
+                        .send(SseStreamManageAsyncCmd::SendMessage(
+                            session_id.clone(),
+                            message,
+                        ))
+                        .await
+                        .ok();
+                }
+                Err(_e) => {}
             }
         }
     };
